Only keep bindgen step, move unused logic in lib.rs
